Add  modular jS files
