+++ src/load.js
@@ -1,9 +1,19 @@
-import display from './display';
-import managerArr from './manager';
+import {getScores, postScore} from './api';
 
-const arr = [];
+
+const user = document.querySelector('#name');
+const score = document.querySelector('#score');
+
+const formButton = document.querySelector('#button-submit');
+const refresh = document.querySelector('.refresh');
+
 const loadEventListeners = () => {
-  window.addEventListener('DOMContentLoaded', managerArr(arr, display));
+  window.addEventListener('DOMContentLoaded', getScores());
+  formButton.onclick = function(){
+    postScore(user, score);
+  } 
+  
+  refresh.addEventListener("click", getScores)
 };
 
 export { loadEventListeners as default };
